refactor: break down position_manager.py into multi-file module
- Split 2098-line monolithic file into 8 logical components
- Created modular structure with clear separation of concerns:
  - core.py: Main PositionManager class and initialization
  - tracking.py: Real-time position tracking and callbacks
  - analytics.py: P&L calculations and portfolio analytics
  - risk.py: Risk metrics and position sizing
  - monitoring.py: Position monitoring and alerts
  - operations.py: Direct position operations (close, partial close)
  - reporting.py: Statistics, history, and report generation
  - types.py: Protocol for proper typing across mixins
- Maintained backward compatibility with existing imports
- Fixed all linting errors and type annotations
- All existing tests continue to pass
