+++ src/project_x_py/position_manager/__init__.py
@@ -1,0 +1,15 @@
+"""
+Position Manager Module for ProjectX Trading Platform.
+
+This module provides comprehensive position management functionality including:
+- Real-time position tracking and monitoring
+- P&L calculations and portfolio analytics
+- Risk metrics and position sizing
+- Position monitoring and alerts
+- Direct position operations (close, partial close)
+- Statistics, history, and report generation
+"""
+
+from .core import PositionManager
+
+__all__ = ["PositionManager"]
